feat: [ASL-4614] Rename and refactor
